stop looking for the plasma version in the session file
this was probably left over from when the module was outside plasma. now
that it is in kinfocenter, and kinfocenter is part of plasma, we can
simply use our own version as the plasma entry version. they are in sync
across all of plasma anyway and because of that kinfocenter's version is
as good an indication for the version of Plasma, the product, as the
session file installed by plasma-workspace

BUG: 436954