add a search field
unlike the filter field this moves through the text to match selection
areas and highlight them, whereas the filter field actually reduces the
text down to only matching lines. filter is like `grep`, search is more
like `less`+searching.

since textarea+scrollview manages focus automatically this is fairly
straight forward. when the search text changes we search for all matches
and collect them in an area of selection areas. moving through the
matches then simply becomes a matter of stepping through the list of
areas.

search is contextually case sensitive. that is: if the needle contains a
capital letter the search is case sensitive otherwise it is not. e.g.
'Present' is sensitive, 'present' is not. this is largely because I'm
not convinced we need a button toggle for the use cases at hand, so I'd
wait with adding one until a user asks for one. for most scenarios
contextual sensitivity should do the job nicely