move cli kcms from textarea to label
as it turns out there are huge performance bottlenecks with textareas
that make it utterly unsuitable for large amounts of text as generate by
some of the cli tools (notably I have a prototype using xdpyinfo and
that takes a solid second to lay out - on load and everytime on window
resize) resulting in lockups of the entire application while the new
document layout is calculated (even when practically the layout
shouldn't change, such as when NoWrap+PlainText is set).

it's very unfortunate because not using textarea means giving up on
selective text copy, but that isn't really any worse than what the
original KCMs had either. so, it is no regression at least.

Label meanwhile performs reasonably well even with the largest payloads
I have at hand, so for the most part it should be a suitable replacement
until

https://bugreports.qt.io/browse/QTBUG-37872
https://bugreports.qt.io/browse/QTBUG-14077