add support for a copy all contextmenu
with the move from textarea to label we've also lost the rather
interesting feature of being able to selectively copy text out of the
cli output kcms.

to mitigate that loss somewhat there now is a context menu on over the
label that allows copying everything currently displayed. this isn't
quite as nice but for many use cases this should still be good enough
and it certainly is better than what the original kcms had.