kcms/energy: More stable UI for history graph
For devices that have history data available, always show the graph.
If there are no data points for the selected time range, simply show
a placeholder message instead of hiding it.

On the other hand, always hide the graph and options for those devices
that cannot provide a history, removing also the potentially confusing
warning message

This moves away from having an annoyingly jumping an unclear UI.

BUG: 490239