kcms/energy: Improve layout of the device selector
Port the device selector to an actual ListView instead of Row+Repeater
combo, so we can use the implicit height of the delegate, instead of
imposing a fixed height, which fails for bigger fonts.

While at it, fix some layout issues within the delegates and simplify
the code by using the ListView regular properties to initialize and
select the current device.

As an added benefit there is now keyboard navigation

BUG: 499290
FIXED-IN: 6.4