kcms/energy: Remove extra top/right paddings and fix plot area
Previous drawing assumed a centered plot with symmetric x and y
paddings, which resulted in extra wasted space on the top and right
of the plot.

Rework the plot positioning to allow different paddings on every
edge, which will serve to accomodate the axis labels.

This new plot geometry property is also used to better center the
placeholder message.

Also limit the min/max graph height so it doesn't get too big on
maximized pages, or too small on compact ones.

BUG: 443569
FIXED-IN: 6.4