kcms/about-distro: Fix hint expanding height of parent layout Instead, use an label with padding to match the height of normal text. A background is used to fill this padded area.
If smallFont is larger than normal text, the text will draw outside the label which would be restricted to the same height as normal text by negative padding. I don't see this as worth fixing - if the small font is larger than normal text, this would be the problem, not here...

When valueLabel spans multiple lines, it will be vertically centered.

CCBUG: 500355